feat: Refactor PTRS routing and enhance Step1 and Step2 components with collapsible table functionality; improve navigation and guidance
